refactor(money): hold amounts as BigDecimal with an explicit currency
Money was a double everywhere, rounded via String.format("%.2f") in two
separate layers, with no currency field anywhere in the platform. This is
defect 9 in the lifecycle spec.

Binary floating point cannot represent most decimal fractions, so a price
held as a double is already not the number that was entered, and the error
compounds through every total derived from it. Amounts are now BigDecimal
at scale 2, rounded HALF_UP, with the contract in orders-api's Money.

Rounding now happens once per amount, where that amount is finalised. A
discounted unit price is rounded as it is set, because it is published on
the order line; the total is the exact sum of those. The old path rounded a
total in ApplyPromotionsService that PricingService then recomputed and
rounded again from unrounded unit prices, so the first rounding was
discarded and the surviving one started from figures that had never been
rounded. That is now observable: 33.33% off 10.00 over qty 3 comes to 20.01,
the sum of the three prices the buyer was shown, where before it was 20.00.

Currency is an explicit field on the order and on amounts crossing a
service boundary, defaulting to EUR. There is one currency and no
conversion; the field exists so amounts are not silently currency-less.
Percentages stay Double - a discount rate is not an amount.

Storage follows the type: product.price and fact_sales_line's unit_price
and line_revenue become numeric(12,2), Mongo stores Decimal128, and Redis
caches BigDecimal. The analytics migration is in this change rather than a
follow-up, because leaving the warehouse on DOUBLE PRECISION would let it
drift from the operational store it is meant to mirror.

The Drools rule needed rewriting: `$product.getPrice() * 0.95` does not
compile against BigDecimal, and the replacement builds its literal from a
string so the binary approximation of 0.95 is not carried into the price.

PricingMoneyTest covers the guarantees rather than the arithmetic: rounded
once, exact at cent scale, denominated, and free of the drift ten additions
of 0.10 produce as doubles.

+++ backend/analytics-service/src/main/java/the/chak/ecommerce/analytics/entity/FactSalesLine.java
@@ -1,5 +1,6 @@
 package the.chak.ecommerce.analytics.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -46,13 +47,13 @@ public class FactSalesLine {
     private Integer units;
 
     @Column(nullable = false)
-    private Double unitPrice;
+    private BigDecimal unitPrice;
 
     /** Discount percentage that applied at purchase time, if any. */
     private Double percentageOff;
 
     @Column(nullable = false)
-    private Double lineRevenue;
+    private BigDecimal lineRevenue;
 
     @Column(nullable = false)
     private LocalDateTime orderDate;
