+++ backend/analytics-service/src/main/java/the/chak/ecommerce/analytics/repository/ProductAggregate.java
@@ -1,5 +1,7 @@
 package the.chak.ecommerce.analytics.repository;
 
+import java.math.BigDecimal;
+
 /** Units sold and revenue for a single product, across all completed orders. */
-public record ProductAggregate(String productId, String title, long units, double revenue) {
+public record ProductAggregate(String productId, String title, long units, BigDecimal revenue) {
 }
